Moved file updates
